Project forest stocks over a plot file

Project forest stocks over a plot file

forecast_biomass(
  save_to = NULL,
  data = NULL,
  plot_id = NULL,
  years = 50,
  clear_start = FALSE,
  clear_year = NULL,
  planting_init = NULL,
  planting_post = NULL,
  allow_colonization = TRUE,
  minimal_if_clearcut = TRUE,
  m_model = NULL,
  u_model = NULL,
  r_model = NULL,
  output_folder_name = "simulation_output"
)

Arguments

save_to

Path to output folder

data

A data frame or CSV file containing the input data in the "matrix model" format.

plot_id

Character. The identifier of the target plot.

years

Numeric. Number of years to forecast.

clear_start

Logical. Starts from clear cut. Defaults to FALSE.

clear_year

Numeric. A year to add a clearcut amid simulation. Defaults to NULL (no clearcut).

planting_init

A data frame with SPCD (species group) and TPH (trees per hectare) columns. This is used to simulate growth of a new plantation at the start if clear_start = TRUE. Default is NULL (no plantation).

planting_post

A data frame with SPCD (species group) and TPH (trees per hectare) columns. This is used to simulate growth of a new plantation after a mid-run clearcut (i.e., when clear_year is procided). Default is NULL (no plantation).

allow_colonization

Logical. Whether to allow other species to colonize the stand. Defaults to FALSE (recruits only from the pre-cut species pool.)

minimal_if_clearcut

Logical.if any clear-cut happens, only save summary and year_summary.

m_model

Path to the mortality model '.rds' file.

u_model

Path to the upgrowth model '.rds' file.

r_model

Path to the recruitment model '.rds' file.

output_folder_name

Character. Name for the output file.

Value

A list with data frames

Examples

if (FALSE) { # \dontrun{
forecast_biomass(data = df, plot_id = "1234")
} # }