Forecast forest stocks for many stands simultaneusly

Forecast forest stocks for many stands simultaneusly

forecast_biomass_batch(
  save_to = NULL,
  data_source = NULL,
  years = 50,
  clear_start = FALSE,
  clear_year = NULL,
  planting_init = NULL,
  planting_post = NULL,
  allow_colonization = TRUE,
  minimal_if_clearcut = TRUE,
  output_folder_name = "forecast_batch_results",
  n_cores = -1
)

Arguments

save_to

Path to output folder

data_source

A data frame, a list of data frames, a CSV file path, or the path to a directory with multiple CSVs (individual plots).

years

Numeric. Number of years to simulate.

clear_start

Logical. Clear at first year. Defaults to FALSE.

clear_year

Numeric. A year to add a clearcut amid simulation. Defaults to NULL (no clearcut).

planting_init

A data frame with SPCD (species group) and TPH (trees per hectare) columns. This is used to simulate growth of a new plantation at the start if clear_start = TRUE. Default is NULL (no plantation).

planting_post

A data frame with SPCD (species group) and TPH (trees per hectare) columns. This is used to simulate growth of a new plantation after a mid-run clearcut (i.e., when clear_year is procided). Default is NULL (no plantation).

allow_colonization

Logical. Whether to allow other species to colonize the stand. Defaults to FALSE (recruits only from the pre-cut species pool.)

minimal_if_clearcut

Logical.if any clear-cut happens, only save summary and year_summary.

output_folder_name

Character. A name for the output folder to be created in the 'save_to' directory.

n_cores

Numeric. Number of cores to use in the parallel process. Default is -1 (all but one).

Value

A list with data frames

Examples

if (FALSE) { # \dontrun{
forecast_biomass_batch(data = df_all)
} # }