Project forest stocks for many plots

DEPRECATED. This function is deprecated. Please use `forecast_biomass()` instead.

project_biomass_batch(
  save_to = NULL,
  data_source = NULL,
  years = 50,
  m_model = NULL,
  u_model = NULL,
  r_model = NULL,
  output_folder_name = "simulation_results_batch",
  n_cores = -1
)

Arguments

save_to

Path to output folder

data_source

A data frame, a list of data frames, a CSV file path, or the path to a directory with multiple CSVs (individual plots).

years

Numeric. Number of years to simulate.

m_model

Path to the mortality model '.rds' file.

u_model

Path to the upgrowth model '.rds' file.

r_model

Path to the recruitment model '.rds' file.

output_folder_name

Character. A name for the output folder to be created in the 'save_to' directory.

n_cores

Numeric. Number of cores to use in the parallel process. Default is -1 (all but one).

Value

A list with data frames

Examples

if (FALSE) { # \dontrun{
project_biomass_batch(data = df_all)
} # }